# backend/specification/api.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: Traffic flow backend specification
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/lawrence9104265/Sharingan/1.0.0
  - url: 'https://petstore.swagger.io/v2'
...
paths:
...
  /Task/Initialize/{mode}:
    post:
      tags:
        - Task
      summary: Initiate a task & Get the task id
      operationId: init_task
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                video_id:
                  type: string
                  default: "some-uuid"
                stabilization:
                  type: number
                  default: 30
                detector:
                  $ref: '#/components/schemas/Detector'
      parameters:
        - name: mode
          in: path
          description: The mode of task
          required: true
          schema:
            type: string
            enum:
              - straight
              - cross_intersection
              - t_intersection
              
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    default: '1234'
...
components:
  schemas:
    Line:
      properties:
        x1:
          type: number
          default: 0
        y1:
          type: number
          default: 0
        x2:
          type: number
          default: 10
        y2:
          type: number
          default: 10
    Detector:
      properties:
        A:
          $ref: '#/components/schemas/Line'
        B:
          $ref: '#/components/schemas/Line'
        X:
          $ref: '#/components/schemas/Line'
        Y:
          $ref: '#/components/schemas/Line'
        T:
          $ref: '#/components/schemas/Line'
...